Remove external ROM emulator connector
Removed the connector for connecting to ROM emulator. The connector was not a particular success as a 34-pin ribbon cable instead of 28 pin is required. The absence of a reset pin was also a mistake. Next generation of HEC will move towards probe-based prototyping or in-situ flashing. Removed R1 and C36 as their role is duplicated.